changing the files order

+++ 2-do_deploy_web_static.py
@@ -37,6 +37,13 @@ def do_deploy(archive_path):
     if removed.failed:
         return False
 
+    movinf_file = run(
+        """mv /data/web_static/releases/{}/web_static/*
+        /data/web_static/releases/{}""".format(
+            file_out_ext[0], file_out_ext[0]))
+    if movinf_file.failed:
+        return False
+
     link_rm = run("rm -rf /data/web_static/current")
     if link_rm.failed:
         return False
